Finished routing
Finished routing the cards to the pages, and passing all the information to the cards and pages.
