--- src/Components/DemoPage.js
@@ -1,10 +1,0 @@
-import React from "react"
-
-const DemoPage = ({name}) => {
-   
-    return(
-<div>{name}</div>
-    )
-}
-
-export default DemoPage
